Move field rewrite into its own floating-chat demo.
Keep the original floating-chat page as a minimal layout example and put the MCP rewrite flow on a separate demo.
