+++ demo/index.html
@@ -118,7 +118,11 @@
       </a>
       <a class="card" href="./load-sidebar-v2-floating-chat/">
         <strong>loadSidebarV2 — floating chat</strong>
-        <span>Floating chat from a text-area right-click; starters and a close-only header.</span>
+        <span>Minimal floating chat widget layout.</span>
+      </a>
+      <a class="card" href="./load-sidebar-v2-floating-chat-rewrite/">
+        <strong>loadSidebarV2 — floating chat rewrite</strong>
+        <span>Right-click a text area to rewrite it with a host MCP server.</span>
       </a>
       <a class="card" href="./load-sidebar-v1-demo/">
         <strong>loadSidebar (v1)</strong>
