+++ demo/load-sidebar-v2-floating-chat-rewrite/ajv-formats-shim.js
